fix: miniwindow.cpp behavioral parity — H25 H30 H75 H76 H78 H79 L85 L88 M130
- H25: DrawImage srcRight/srcBottom now apply FixRight/FixBottom (<=0 →
  offset from right/bottom edge) matching original miniwindow.cpp:1392-1396;
  replaces the all-zeros fallback that missed negative values.
- H30: BlendImage rejects opacity outside [0.0,1.0] with eBadParameter
  instead of silently clamping (original: miniwindow.cpp:2067-2068).
- H75 (MEDIUM): Resize no longer overwrites backgroundColor; original
  only updates width/height (miniwindow.cpp:4166-4167).
- H75 (LOW): Resize coerces zero/negative dimensions to 1 via MAX(w,1)
  instead of returning eBadParameter (miniwindow.cpp:4149).
- H76 (MEDIUM): TextWidth returns -2 when font not found
  (original: miniwindow.cpp:894-895).
- H76 (LOW): Text returns 0 immediately for empty string without drawing
  or setting the dirty flag (original: miniwindow.cpp:833-834).
- H78/H79: CircleOp pattern brushes (styles 8-12) now use correct color
  semantics: set-bits=brushColor, clear-bits=penColor (original SetTextColor/
  SetBkColor assignments). brushColor==-1 skips background fill (TRANSPARENT
  mode). Styles 11/12 now render as wave bitmap patterns instead of solid fill,
  matching the original 8×8 monochrome bitmaps.
- L85: WindowGetPixel Lua binding casts QRgb to qint32 before pushnumber so
  out-of-bounds returns -1 in Lua (not 4294967295) matching the original long
  return type (miniwindow.cpp:3306-3308).
- L88: Arc endpoint (x2,y2) now applies FixRight/FixBottom so negative values
  are treated as offsets from the right/bottom edge (original: line 991).
- M130: Gradient invalid mode returns eUnknownOption instead of eBadParameter
  (original: miniwindow.cpp:2678,2738).

Adds 7 GoogleTest cases in test_miniwindow_loadimage_gtest.cpp covering H30,
H75×2, H76×2, M130, and GetPixel signed-cast. Build: clean. Tests: 17/17 pass.

+++ src/world/lua_api/world_miniwindow_drawing.cpp
@@ -405,7 +405,10 @@ int L_WindowGetPixel(lua_State* L)
     }
 
     QRgb color = win->GetPixel(x, y);
-    lua_pushnumber(L, color);
+    // Original CMiniWindow::GetPixel returns long (signed 32-bit). CLR_INVALID=0xFFFFFFFF
+    // becomes -1 when cast to long. Push as signed so Lua scripts see -1 for out-of-bounds,
+    // not 4294967295 (original: miniwindow.cpp:3306-3308).
+    lua_pushnumber(L, static_cast<qint32>(color));
     return 1;
 }
